fuctions to encodingUtils
