+++ src/openAIFunctions.ts
@@ -2,7 +2,8 @@ import OpenAI from 'openai';
 import { MyContext, MyMessage, MyMessageContent, UserData } from './types';
 import { AxiosResponse } from 'axios';
 import pTimeout from 'p-timeout';
-import { toLogFormat, encodeText, decodeTokens } from './utils';
+import { toLogFormat } from './utils';
+import { encodeText, decodeTokens } from './encodingUtils';
 import { timeoutMsDefaultchatGPT, GPT_MODEL, maxTokensThresholdToReduceHistory, defaultPromptMessage } from './config';
 import { usedTokensForUser, insertUserOrUpdate, selectUserByUserId } from './database';
 import { maxTrialsTokens, OPENAI_API_KEY } from './config';
